+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/FlavourContext.java
@@ -30,4 +30,23 @@ public class FlavourContext {
     @Getter @Setter
     private List<String> modules;
 
+    @Getter @Setter
+    private boolean addLombok;
+
+    private String toClassName( String base, String splitString ) {
+        StringBuilder buf = new StringBuilder();
+        String[] split = base.split( splitString );
+        for ( String part : split ) {
+            if ( part.length() > 0 ) {
+                buf.append( part.substring( 0, 1 ).toLowerCase()+part.substring( 1 ) );
+            }
+        }
+        return buf.toString();
+    }
+
+    public String getArtifactIdAsClassName() {
+        String res = toClassName( this.artifactId, "-" );
+        return toClassName( res, "\\." );
+    }
+
 }