+++ fj-doc-playground-quarkus/src/main/react/src/playground/DocProjectInit.jsx
@@ -212,6 +212,7 @@ const DocProjectInit = () => {
 							<MenuItem value={'vanilla'}>Vanilla (Simple library project)</MenuItem>
 							<MenuItem value={'quarkus-3'}>Quarkus 3 (Recommended quarkus version)</MenuItem>
 							<MenuItem value={'quarkus-2'}>Quarkus 2 (Legacy quarkus version)</MenuItem>
+							<MenuItem value={'micronaut-4'}>Micronaut 4</MenuItem>
 						</Select>
 					</FormControl>
 				</Grid>